first render of menus
