+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\src\\index.js":"1","C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\src\\App.js":"2","C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\src\\reportWebVitals.js":"3","C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\src\\components\\MenuSupper.js":"4","C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\src\\components\\MenuBreakfast.js":"5","C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\src\\components\\Bar.js":"6"},{"size":500,"mtime":1607090658470,"results":"7","hashOfConfig":"8"},{"size":335,"mtime":1607214901375,"results":"9","hashOfConfig":"8"},{"size":362,"mtime":1607090658412,"results":"10","hashOfConfig":"8"},{"size":523,"mtime":1607398394042,"results":"11","hashOfConfig":"8"},{"size":1054,"mtime":1607399233395,"results":"12","hashOfConfig":"8"},{"size":284,"mtime":1607223619383,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"ra8cpa",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\src\\index.js",[],["27","28"],"C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\src\\App.js",[],"C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\src\\reportWebVitals.js",[],"C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\src\\components\\MenuSupper.js",[],"C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\src\\components\\MenuBreakfast.js",[],"C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\src\\components\\Bar.js",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
