--- burger-queen/.eslintcache
@@ -1,1 +1,0 @@
-[{"C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\burger-queen\\src\\index.js":"1","C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\burger-queen\\src\\App.js":"2","C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\burger-queen\\src\\reportWebVitals.js":"3","C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\burger-queen\\src\\components\\MenuBreakfast.js":"4","C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\burger-queen\\src\\components\\Contador.js":"5","C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\burger-queen\\src\\components\\MenuSupper.js":"6","C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\burger-queen\\src\\components\\Bar.js":"7"},{"size":500,"mtime":1607090658470,"results":"8","hashOfConfig":"9"},{"size":335,"mtime":1607214901375,"results":"10","hashOfConfig":"9"},{"size":362,"mtime":1607090658412,"results":"11","hashOfConfig":"9"},{"size":534,"mtime":1607224398732,"results":"12","hashOfConfig":"9"},{"size":413,"mtime":1607128993246,"results":"13","hashOfConfig":"9"},{"size":526,"mtime":1607224326329,"results":"14","hashOfConfig":"9"},{"size":284,"mtime":1607223619383,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"1o9qyx6",{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"27"},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\burger-queen\\src\\index.js",[],["32","33"],"C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\burger-queen\\src\\App.js",[],"C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\burger-queen\\src\\reportWebVitals.js",[],"C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\burger-queen\\src\\components\\MenuBreakfast.js",[],"C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\burger-queen\\src\\components\\Contador.js",[],["34","35"],"C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\burger-queen\\src\\components\\MenuSupper.js",[],"C:\\Users\\USER\\Documents\\data\\BOG001-burger-queen\\burger-queen\\src\\components\\Bar.js",[],{"ruleId":"36","replacedBy":"37"},{"ruleId":"38","replacedBy":"39"},{"ruleId":"36","replacedBy":"40"},{"ruleId":"38","replacedBy":"41"},"no-native-reassign",["42"],"no-negated-in-lhs",["43"],["42"],["43"],"no-global-assign","no-unsafe-negation"]
