+++ firestore.indexes.json
@@ -1,3 +1,4 @@
+  
 {
   "indexes": []
 }
