+++ src/components/Bar.js
