+++ src/components/lista.js
@@ -1,0 +1,1 @@
+
